+++ src/controller/Admin.php
@@ -29,16 +29,20 @@ class Admin extends AuthController{
     protected $config;
     /**
      * @Inyectable
-     * @var \PSFS\service\AdminService Servicios de administración
+     * @var \PSFS\services\AdminServices Servicios de administración
      */
     protected $srv;
+    /**
+     * @Inyectable
+     * @var  \PSFS\services\GeneratorService Servicio de generación de estructura de directorios
+     */
+    protected $gen;
     /**
      * Constructor por defecto
      */
     public function __construct()
     {
         $this->init();
-        pre($this);
         $this->setDomain('ROOT')
             ->setTemplatePath($this->config->getTemplatePath());
     }
@@ -224,11 +228,11 @@ class Admin extends AuthController{
                 $module = $form->getFieldValue("module");
                 try
                 {
-                    $this->srv->createStructureModule($module, Logger::getInstance());
+                    $this->gen->createStructureModule($module, Logger::getInstance());
                     return $this->getRequest()->redirect(Router::getInstance()->getRoute("admin-module", true));
                 }catch(\Exception $e)
                 {
-                    Logger::getInstance()->infoLog($e->getMessage() . "[" . $e->getLine() . "]");
+                    Logger::getInstance()->infoLog($e->getMessage() . " [" . $e->getFile() . ":" . $e->getLine() . "]");
                     throw new \HttpException('Error al generar el módulo, prueba a cambiar los permisos', 403);
                 }
             }
