+++ src/templates/generator/service.template.twig
@@ -1,0 +1,13 @@
+<?php
+namespace {{ module }}\Services;
+
+use PSFS\base\Service;
+
+/**
+* Class {{ module }}
+* @package {{ module }}\Services
+*/
+class {{ module }}Service extends Service
+{
+
+}
